Added live updating of User's MyGarden.html
MyGarden.html now reflects the user's created plants.

However, similar to the login redirect, it takes about 2 seconds for the javascript to keep up with Firebase's updated repository. So we when revisiting MyGarden.html ,we will need to delay the user, which leads to the plantStall.html.

However, I've made it that when you first add the new plant, it will create a new element locally. Only when you want to revisit MyGardens.html again will you need to redirect.